+++ src/index.ts
@@ -10,7 +10,6 @@ import { FoeSheet } from './module/actor/sheets/foesheet'
 import { StarforgedCharacterSheet } from './module/actor/sheets/sf-charactersheet'
 import { StarforgedLocationSheet } from './module/actor/sheets/sf-locationsheet'
 import { IronswornSharedSheetV2 } from './module/actor/sheets/sharedsheet-v2'
-import { IronswornSiteSheet } from './module/actor/sheets/sitesheet'
 import { IronswornSiteSheetV2 } from './module/actor/sheets/sitesheet-v2'
 import { StarshipSheet } from './module/actor/sheets/starshipsheet'
 import { CreateActorDialog } from './module/applications/createActorDialog'
@@ -108,12 +107,8 @@ Hooks.once('init', async () => {
 
   Actors.registerSheet('ironsworn', IronswornSiteSheetV2, {
     types: ['site'],
-    label: 'Site sheet v2',
+    label: 'Site sheet',
     makeDefault: true,
-  })
-  Actors.registerSheet('ironsworn', IronswornSiteSheet, {
-    types: ['site'],
-    label: 'Classic site sheet',
   })
 
   Items.registerSheet('ironsworn', AssetSheetV2, {