+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove "Oracle NN:" from Ironsworn oracle names ([#453](https://github.com/ben/foundry-ironsworn/pull/453))
 - Fix a visual glitch with progress tracks in v10
+- Retire the classic Site sheet
 
 ## 1.17.2